refactor(st-09054): modularize reflection nodes
