+++ docs/st09054-reflection-node-modularization.md
@@ -1,0 +1,50 @@
+# ST-09054: Reflection Node Modularization
+
+## Summary
+
+`packages/patterns/src/reflection/nodes.ts` was reduced from `350` lines to a `13` line public facade by extracting focused internal modules for the main reflection-node responsibilities:
+
+- `packages/patterns/src/reflection/node-shared.ts`
+- `packages/patterns/src/reflection/generator-node.ts`
+- `packages/patterns/src/reflection/reflector-node.ts`
+- `packages/patterns/src/reflection/reviser-node.ts`
+- `packages/patterns/src/reflection/finisher-node.ts`
+
+The facade still owns the stable reflection-node export surface, while generation, reflection, revision, finish-state completion, and shared content/history utilities now live behind explicit internal seams. Every extracted production module stays below the `300` line cutoff, with the largest extracted file at `133` lines.
+
+## Test Modularization
+
+The monolithic `packages/patterns/tests/reflection/nodes.test.ts` suite was replaced with focused reflection-node behavior suites:
+
+- `packages/patterns/tests/reflection/generator-node.test.ts`
+- `packages/patterns/tests/reflection/reflector-node.test.ts`
+- `packages/patterns/tests/reflection/reviser-node.test.ts`
+- `packages/patterns/tests/reflection/finisher-node.test.ts`
+
+This keeps generator, reflector, reviser, and finisher behavior independently reviewable instead of coupling all reflection-node assertions to one file.
+
+## Compatibility Notes
+
+- Public reflection-node exports remain available through `packages/patterns/src/reflection/nodes.ts`, `packages/patterns/src/reflection/index.ts`, and the top-level patterns package exports.
+- Existing reflection-node behavior, iteration semantics, logging names, and finish-state behavior remain preserved behind the split facade.
+- Reflection parsing still falls back to a plain-text critique entry when the model response is not valid JSON.
+- Revision-history prompt context stays shared between reflector and reviser paths without changing the visible prompt templates or revision truncation behavior.
+
+## Validation
+
+- `pnpm test --run packages/patterns/tests/reflection/generator-node.test.ts packages/patterns/tests/reflection/reflector-node.test.ts packages/patterns/tests/reflection/reviser-node.test.ts packages/patterns/tests/reflection/finisher-node.test.ts`
+  - `4` files passed, `12` tests passed
+- `pnpm --filter @agentforge/patterns typecheck`
+- `pnpm --filter @agentforge/patterns exec eslint src/reflection/nodes.ts src/reflection/node-shared.ts src/reflection/generator-node.ts src/reflection/reflector-node.ts src/reflection/reviser-node.ts src/reflection/finisher-node.ts tests/reflection/generator-node.test.ts tests/reflection/reflector-node.test.ts tests/reflection/reviser-node.test.ts tests/reflection/finisher-node.test.ts`
+- `pnpm lint:explicit-any:baseline`
+  - workspace: `84/289`
+  - patterns: `2/28`
+
+## Explicit-`any` Notes
+
+- This story did not increase the explicit-`any` baseline.
+- The touched reflection-node modules do not introduce new explicit-`any` warnings.
+
+## CI Impact
+
+- No CI workflow change was required because the existing patterns typecheck, lint, explicit-`any`, and full-suite test coverage already exercises the reflection runtime and package build surface after the modularization split.
