+++ packages/patterns/src/reflection/nodes.ts
@@ -1,350 +1,13 @@
 /**
  * Node Implementations for Reflection Pattern
  *
- * This module implements the core nodes for the Reflection agent pattern.
+ * This module keeps the stable public reflection-node export surface while
+ * delegating each node implementation to a focused internal module.
  *
  * @module patterns/reflection/nodes
  */
 
-import { HumanMessage, SystemMessage } from '@langchain/core/messages';
-import type { ReflectionStateType } from './state.js';
-import type { GeneratorConfig, ReflectorConfig, ReviserConfig } from './types.js';
-import type { Reflection } from './schemas.js';
-import {
-  DEFAULT_GENERATOR_SYSTEM_PROMPT,
-  DEFAULT_REFLECTOR_SYSTEM_PROMPT,
-  DEFAULT_REVISER_SYSTEM_PROMPT,
-  GENERATION_PROMPT_TEMPLATE,
-  REFLECTION_PROMPT_TEMPLATE,
-  REVISION_PROMPT_TEMPLATE,
-  QUALITY_CRITERIA_TEMPLATE,
-  REFLECTION_HISTORY_TEMPLATE,
-  REFLECTION_ENTRY_TEMPLATE,
-  REVISION_ENTRY_TEMPLATE,
-} from './prompts.js';
-import { createPatternLogger } from '../shared/deduplication.js';
-import { handleNodeError } from '../shared/error-handling.js';
-
-// Create loggers for Reflection pattern nodes
-const generatorLogger = createPatternLogger('agentforge:patterns:reflection:generator');
-const reflectorLogger = createPatternLogger('agentforge:patterns:reflection:reflector');
-const reviserLogger = createPatternLogger('agentforge:patterns:reflection:reviser');
-
-/**
- * Create a generator node that creates initial responses
- */
-export function createGeneratorNode(config: GeneratorConfig) {
-  const {
-    model,
-    systemPrompt = DEFAULT_GENERATOR_SYSTEM_PROMPT,
-    verbose = false,
-  } = config;
-
-  return async (state: ReflectionStateType): Promise<Partial<ReflectionStateType>> => {
-    const startTime = Date.now();
-
-    try {
-      generatorLogger.debug('Generating response', {
-        attempt: state.iteration + 1,
-        hasFeedback: state.reflections.length > 0,
-        hasExistingResponse: !!state.currentResponse
-      });
-
-      // Build context from previous iterations if any
-      let context = '';
-      if (state.iteration > 0 && state.reflections.length > 0) {
-        const lastReflection = state.reflections[state.reflections.length - 1];
-        context = `\nPrevious feedback to consider:\n${lastReflection.critique}`;
-      }
-
-      const userPrompt = GENERATION_PROMPT_TEMPLATE
-        .replace('{input}', state.input || '')
-        .replace('{context}', context);
-
-      // Call LLM to generate response
-      const messages = [
-        new SystemMessage(systemPrompt),
-        new HumanMessage(userPrompt),
-      ];
-
-      const response = await model.invoke(messages);
-      const content = typeof response.content === 'string'
-        ? response.content
-        : JSON.stringify(response.content);
-
-      generatorLogger.info('Response generated', {
-        attempt: state.iteration + 1,
-        responseLength: content.length,
-        isRevision: state.iteration > 0,
-        duration: Date.now() - startTime
-      });
-
-      return {
-        currentResponse: content,
-        status: 'reflecting' as const,
-        iteration: 1,
-      };
-    } catch (error) {
-      // Handle error with proper GraphInterrupt detection
-      const errorMessage = handleNodeError(error, 'generator', false);
-
-      generatorLogger.error('Response generation failed', {
-        attempt: state.iteration + 1,
-        error: errorMessage,
-        duration: Date.now() - startTime
-      });
-      return {
-        status: 'failed' as const,
-        error: errorMessage,
-      };
-    }
-  };
-}
-
-/**
- * Create a reflector node that critiques responses
- */
-export function createReflectorNode(config: ReflectorConfig) {
-  const {
-    model,
-    systemPrompt = DEFAULT_REFLECTOR_SYSTEM_PROMPT,
-    qualityCriteria,
-    verbose = false,
-  } = config;
-
-  return async (state: ReflectionStateType): Promise<Partial<ReflectionStateType>> => {
-    const startTime = Date.now();
-
-    try {
-      reflectorLogger.debug('Reflecting on response', {
-        attempt: state.iteration,
-        responseLength: state.currentResponse?.length || 0,
-        hasCriteria: !!(qualityCriteria || state.qualityCriteria)
-      });
-
-      if (!state.currentResponse) {
-        throw new Error('No current response to reflect on');
-      }
-
-      // Build quality criteria section
-      let criteriaSection = '';
-      const criteria = qualityCriteria || state.qualityCriteria;
-      if (criteria) {
-        const criteriaList = criteria.criteria?.join('\n- ') || '';
-        criteriaSection = QUALITY_CRITERIA_TEMPLATE
-          .replace('{criteria}', criteriaList ? `- ${criteriaList}` : 'General quality standards')
-          .replace('{minScore}', criteria.minScore?.toString() || '7')
-          .replace('{requireAll}', criteria.requireAll ? 'All criteria must be met.' : 'Meet as many criteria as possible.');
-      }
-
-      // Build history section
-      let historySection = '';
-      if (state.reflections.length > 0 || state.revisions.length > 0) {
-        const reflectionsText = state.reflections
-          .map((r, idx) => REFLECTION_ENTRY_TEMPLATE
-            .replace('{iteration}', (idx + 1).toString())
-            .replace('{critique}', r.critique)
-            .replace('{score}', r.score?.toString() || 'N/A')
-            .replace('{issues}', r.issues.join(', '))
-            .replace('{suggestions}', r.suggestions.join(', '))
-          )
-          .join('\n\n');
-
-        const revisionsText = state.revisions
-          .map((r) => REVISION_ENTRY_TEMPLATE
-            .replace('{iteration}', r.iteration.toString())
-            .replace('{content}', r.content.substring(0, 200) + '...')
-          )
-          .join('\n\n');
-
-        historySection = REFLECTION_HISTORY_TEMPLATE
-          .replace('{reflections}', reflectionsText || 'None')
-          .replace('{revisions}', revisionsText || 'None');
-      }
-
-      const userPrompt = REFLECTION_PROMPT_TEMPLATE
-        .replace('{input}', state.input || '')
-        .replace('{currentResponse}', state.currentResponse)
-        .replace('{criteria}', criteriaSection)
-        .replace('{history}', historySection);
-
-      // Call LLM to generate reflection
-      const messages = [
-        new SystemMessage(systemPrompt),
-        new HumanMessage(userPrompt),
-      ];
-
-      const response = await model.invoke(messages);
-      const content = typeof response.content === 'string'
-        ? response.content
-        : JSON.stringify(response.content);
-
-      // Parse the reflection from JSON
-      let reflection: Reflection;
-      try {
-        // Try to extract JSON from the response
-        const jsonMatch = content.match(/\{[\s\S]*\}/);
-        if (jsonMatch) {
-          reflection = JSON.parse(jsonMatch[0]);
-        } else {
-          // Fallback: create a basic reflection
-          reflection = {
-            critique: content,
-            issues: [],
-            suggestions: [],
-            score: 5,
-            meetsStandards: false,
-          };
-        }
-      } catch (parseError) {
-        // Fallback: create a basic reflection
-        reflection = {
-          critique: content,
-          issues: [],
-          suggestions: [],
-          score: 5,
-          meetsStandards: false,
-        };
-      }
-
-      reflectorLogger.info('Reflection complete', {
-        attempt: state.iteration,
-        ...(reflection.score !== undefined ? { score: reflection.score } : {}),
-        meetsStandards: reflection.meetsStandards,
-        issueCount: reflection.issues.length,
-        suggestionCount: reflection.suggestions.length,
-        duration: Date.now() - startTime
-      });
-
-      return {
-        reflections: [reflection],
-        status: reflection.meetsStandards ? 'completed' as const : 'revising' as const,
-      };
-    } catch (error) {
-      // Handle error with proper GraphInterrupt detection
-      const errorMessage = handleNodeError(error, 'reflector', false);
-
-      reflectorLogger.error('Reflection failed', {
-        attempt: state.iteration,
-        error: errorMessage,
-        duration: Date.now() - startTime
-      });
-      return {
-        status: 'failed' as const,
-        error: errorMessage,
-      };
-    }
-  };
-}
-
-/**
- * Create a reviser node that improves responses based on critiques
- */
-export function createReviserNode(config: ReviserConfig) {
-  const {
-    model,
-    systemPrompt = DEFAULT_REVISER_SYSTEM_PROMPT,
-    verbose = false,
-  } = config;
-
-  return async (state: ReflectionStateType): Promise<Partial<ReflectionStateType>> => {
-    const startTime = Date.now();
-
-    try {
-      if (!state.currentResponse) {
-        throw new Error('No current response to revise');
-      }
-
-      if (state.reflections.length === 0) {
-        throw new Error('No reflections to base revision on');
-      }
-
-      const lastReflection = state.reflections[state.reflections.length - 1];
-
-      reviserLogger.debug('Revising response', {
-        attempt: state.iteration,
-        ...(lastReflection.score !== undefined ? { previousScore: lastReflection.score } : {}),
-        issueCount: lastReflection.issues.length,
-        suggestionCount: lastReflection.suggestions.length
-      });
-
-      // Build history section
-      let historySection = '';
-      if (state.revisions.length > 0) {
-        const revisionsText = state.revisions
-          .map((r) => REVISION_ENTRY_TEMPLATE
-            .replace('{iteration}', r.iteration.toString())
-            .replace('{content}', r.content.substring(0, 200) + '...')
-          )
-          .join('\n\n');
-
-        historySection = `\nPrevious Revisions:\n${revisionsText}`;
-      }
-
-      const userPrompt = REVISION_PROMPT_TEMPLATE
-        .replace('{input}', state.input || '')
-        .replace('{currentResponse}', state.currentResponse)
-        .replace('{critique}', lastReflection.critique)
-        .replace('{issues}', lastReflection.issues.join('\n- '))
-        .replace('{suggestions}', lastReflection.suggestions.join('\n- '))
-        .replace('{history}', historySection);
-
-      // Call LLM to generate revision
-      const messages = [
-        new SystemMessage(systemPrompt),
-        new HumanMessage(userPrompt),
-      ];
-
-      const response = await model.invoke(messages);
-      const content = typeof response.content === 'string'
-        ? response.content
-        : JSON.stringify(response.content);
-
-      // Create revision entry
-      const revision = {
-        content,
-        iteration: state.iteration,
-        basedOn: lastReflection,
-      };
-
-      reviserLogger.info('Revision complete', {
-        attempt: state.iteration,
-        revisionLength: content.length,
-        ...(lastReflection.score !== undefined ? { basedOnScore: lastReflection.score } : {}),
-        duration: Date.now() - startTime
-      });
-
-      return {
-        currentResponse: content,
-        revisions: [revision],
-        status: 'reflecting' as const,
-        iteration: 1,
-      };
-    } catch (error) {
-      // Handle error with proper GraphInterrupt detection
-      const errorMessage = handleNodeError(error, 'reviser', false);
-
-      reviserLogger.error('Revision failed', {
-        attempt: state.iteration,
-        error: errorMessage,
-        duration: Date.now() - startTime
-      });
-      return {
-        status: 'failed' as const,
-        error: errorMessage,
-      };
-    }
-  };
-}
-
-/**
- * Create a finisher node that marks the reflection as complete
- */
-export function createFinisherNode() {
-  return async (state: ReflectionStateType): Promise<Partial<ReflectionStateType>> => {
-    return {
-      status: 'completed' as const,
-      response: state.currentResponse,
-    };
-  };
-}
+export { createGeneratorNode } from './generator-node.js';
+export { createReflectorNode } from './reflector-node.js';
+export { createReviserNode } from './reviser-node.js';
+export { createFinisherNode } from './finisher-node.js';
