+++ planning/features/09-solid-micro-refactors-feature-plan.md
@@ -2,8 +2,8 @@
 
 **Epic Range:** EP-09 through EP-09
 **Status:** In Progress
-**Last Updated:** 2026-05-28
-**Active Story:** None
+**Last Updated:** 2026-05-29
+**Active Story:** ST-09054
 
 ---
